Move cell to board
Clean up remaining point refences and removed imports to UIKit
