--- TicTacToeCheatTool/src/TicTacToeCell.swift
@@ -1,19 +1,0 @@
-//
-//  TicTacToeCell.swift
-//  TicTacToeCheatTool
-//
-//  Created by Nathan Christensen on 2018-05-30.
-//  Copyright © 2018 Nathan Christensen. All rights reserved.
-//
-
-import Foundation
-
-struct TicTacToeCell {
-    var row: Int
-    var column: Int
-    
-    init(row: Int, column: Int) {
-        self.row = row
-        self.column = column
-    }
-}
